+++ src/styles/Films.module.css
@@ -80,6 +80,7 @@
   .filmCard {
     width: 31%;
     margin: 0.5rem;
+    box-shadow: 8px 4px 16px rgba(0, 0, 0, 0.4);
   }
 
   .title {
